+++ Sources/FKUIKit/Components/Button/Public/FKButton/Interaction/FKButton+InterfaceBuilderHooks.swift
@@ -3,6 +3,8 @@ import UIKit
 extension FKButton {
   open override func prepareForInterfaceBuilder() {
     super.prepareForInterfaceBuilder()
-    flushPendingRefresh()
+    MainActor.assumeIsolated {
+      flushPendingRefresh()
+    }
   }
 }